Added DI so I can test input from stdin
After some Googling around, I found this SO post about testing input from stdin: https://stackoverflow.com/questions/28370126/how-can-i-test-stdin-and-stdout, and I decided to rewrite the codebase a little to allow for this testing.

More or less, all I've done here is convert all the functions inside the maybe module into methods on the prompto type.

This has the added benefit of opening up my class to other objects that implement BufRead and Write, so I'm curious to see if I could, for instance, test these functions with a file.