I accidentally my documentation
So, it turns out I've totally misunderstood how Rust's hierarchy plays with documentation. I thought it was enough to have the functions be public, but it turns out I need the struct they're implemented on to be public as well. Long story short, I need to reorganize the API now...